fix: move package for vault sync to avoid collisions in aggregate reports.
