+++ sync/vault-sync/src/main/java/co/com/bancolombia/secretsmanager/vaultsync/connector/VaultAuthenticator.java
@@ -1,12 +1,12 @@
-package co.com.bancolombia.secretsmanager.connector;
+package co.com.bancolombia.secretsmanager.vaultsync.connector;
 
 import co.com.bancolombia.secretsmanager.api.exceptions.SecretException;
 import co.com.bancolombia.secretsmanager.commons.utils.GsonUtils;
-import co.com.bancolombia.secretsmanager.config.VaultSecretsManagerProperties;
-import co.com.bancolombia.secretsmanager.connector.auth.AuthResponse;
-import co.com.bancolombia.secretsmanager.connector.auth.K8sAuth;
-import co.com.bancolombia.secretsmanager.connector.auth.K8sTokenReader;
-import co.com.bancolombia.secretsmanager.connector.auth.RoleAuth;
+import co.com.bancolombia.secretsmanager.vaultsync.config.VaultSecretsManagerProperties;
+import co.com.bancolombia.secretsmanager.vaultsync.connector.auth.AuthResponse;
+import co.com.bancolombia.secretsmanager.vaultsync.connector.auth.K8sAuth;
+import co.com.bancolombia.secretsmanager.vaultsync.connector.auth.K8sTokenReader;
+import co.com.bancolombia.secretsmanager.vaultsync.connector.auth.RoleAuth;
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import com.google.gson.Gson;
