+++ sync/vault-sync/src/main/java/co/com/bancolombia/secretsmanager/vaultsync/connector/auth/K8sAuth.java
@@ -1,4 +1,4 @@
-package co.com.bancolombia.secretsmanager.connector.auth;
+package co.com.bancolombia.secretsmanager.vaultsync.connector.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
