+++ sync/vault-sync/src/main/java/co/com/bancolombia/secretsmanager/vaultsync/connector/ssl/SslConfig.java
@@ -1,4 +1,4 @@
-package co.com.bancolombia.secretsmanager.connector.ssl;
+package co.com.bancolombia.secretsmanager.vaultsync.connector.ssl;
 
 import co.com.bancolombia.secretsmanager.api.exceptions.SecretException;
