+++ sync/vault-sync/src/test/java/co/com/bancolombia/secretsmanager/vaultsync/connector/auth/K8sTokenReaderTest.java
@@ -1,4 +1,4 @@
-package co.com.bancolombia.secretsmanager.connector.auth;
+package co.com.bancolombia.secretsmanager.vaultsync.connector.auth;
 
 import co.com.bancolombia.secretsmanager.api.exceptions.SecretException;
 import lombok.SneakyThrows;
